requirements.txt
fastapi
uvicorn[standard]
streamlit
requests

langchain
langchain-community
langchain-huggingface
langchain-groq
langchain-text-splitters

sentence-transformers
faiss-cpu
transformers
accelerate

pypdf

sqlalchemy

python-dotenv
python-multipart